+++ src/Preprocess/hough_transform.h
@@ -39,8 +39,8 @@ Line *find_image_lines(SDL_Surface *image, int nb_lines, int draw_lines);
 matrice *create_hough_space_matrice(SDL_Surface *image, 
         array *thetas, 
         array *rhos, 
-        float d,
-        float drho);
+        double d,
+        double drho);
 
 
 // finds the top nb_lines peaks in the accumulator matrix.
@@ -61,5 +61,5 @@ void draw_hough_lines_on_image(SDL_Surface *image,
                                int nb_lines,
                                array *thetas, 
                                array *rhos, 
-                               float d);
+                               double d);
