+++ src/Utils/matrice.h
@@ -25,11 +25,11 @@ matrice *matrice_read_csv(char *filename);
 
 matrice *matrice_clone(matrice *m);
 
-float matrice_get(matrice *m, int row, int column);
-float *matrice_get_ref(matrice *m, int row, int column);
-void matrice_set(matrice *m, int row, int column, float value);
+double matrice_get(matrice *m, int row, int column);
+double *matrice_get_ref(matrice *m, int row, int column);
+void matrice_set(matrice *m, int row, int column, double value);
 
-matrice *matrice_random(int rows, int columns, float min, float max);
+matrice *matrice_random(int rows, int columns, double min, double max);
 matrice *matrice_zeros(int rows, int columns);
 
 int matrice_equals(matrice *m1, matrice *m2);
@@ -39,21 +39,21 @@ matrice *matrice_transpose(matrice *m);
 
 // element wise operations
 matrice *matrice_elementwise(matrice *m1, matrice *m2,
-                             float (*f)(float, float));
+                             double (*f)(double, double));
 matrice *matrice_add(matrice *m1, matrice *m2);
 matrice *matrice_sub(matrice *m1, matrice *m2);
 matrice *matrice_mul(matrice *m1, matrice *m2);
 
 // inplace operations
-matrice *matrice_map(matrice *m, float (*f)(float));
-matrice *matrice_multiply(matrice *m, float scalar);
+matrice *matrice_map(matrice *m, double (*f)(double));
+matrice *matrice_multiply(matrice *m, double scalar);
 
 void matrice_add_inplace(matrice *dest, matrice *source);
 void matrice_sub_inplace(matrice *dest, matrice *source);
 
-float *matrice_max(matrice *m, int *row, int *column);
+double *matrice_max(matrice *m, int *row, int *column);
 
-float matrice_sum(matrice *m);
-float matrice_mean(matrice *m);
+double matrice_sum(matrice *m);
+double matrice_mean(matrice *m);
 // std = standard deviation = écart type = sqrt(variance)
-float matrice_std(matrice *m);
+double matrice_std(matrice *m);
